ContractAutomaton.hs: Refactor -- factor Name out. Refactor return type of `contract`
